fix sign warning
git-svn-id: svn://svn.lighttpd.net/xcache/trunk@1207 c26eb9a1-5813-0410-bd6c-c2e55f420ca7

+++ mod_coverager/xc_coverager.c
@@ -183,7 +183,7 @@ static void xc_coverager_save_cov(char *srcfile, char *outfilename, coverager_t 
 		goto bailout;
 	}
 	lseek(fd, 0, SEEK_SET);
-	if (write(fd, (char *) buf, size) != size) {
+	if (write(fd, (char *) buf, size) != (ssize_t) size) {
 		goto bailout;
 	}
